test: create deployment helm chart and e2e test
this helm chart isn't intended for production use. it is far from it as
it gives the pods full access to the api server. further tunning will be
needed to make this a real thing but it is decent enough to at least get
us a running environment where we can test it.

see instruction in the newly created deploying.md file to understand how
this is intended to work.

this commit also introduces our first e2e test.