feat!: remove support for on the fly queue admission
we only operate on queues if they are part of the configuration and if
they aren't we fail. from this point on we don't need to support an
arbitrary number of queues so the feature to add new queues isnt needed
anymore.

this commits removes the feature as to keep the code simpler specially
now we will be introducing new queue reading algorithms.