feat: check bindings len agains max size
if a given placement request bindings' length is bigger than the max
allowed for the scheduler the placement request must be rejected. this
commit also rejects if the placement request comes from an unknown
scheduler (for one we do not have a queue configured).

configuration types were merged using embeddings so we can avoid
duplicating some types.